Abstracting overlay box container from ToastOverlay and LoadingOverlay into OverlayContainer component; Other minor adjustments pointed in PR review.
