+++ cardstack/src/components/LoadingOverlay/LoadingOverlay.tsx
@@ -1,31 +1,12 @@
 import React, { memo } from 'react';
-import { StyleSheet, ActivityIndicator } from 'react-native';
+import { ActivityIndicator } from 'react-native';
 
-import { CenteredContainer, Text } from '@cardstack/components';
+import {
+  OverlayContainer,
+  CenteredContainer,
+  Text,
+} from '@cardstack/components';
 import { colors } from '@cardstack/theme';
-
-const styles = StyleSheet.create({
-  overlayWrapper: {
-    flex: 1,
-    // Note: Workaround for android gesture transparency issue.
-    backgroundColor: '#00000001',
-  },
-});
-
-const overlayStyle = {
-  shadowOffset: {
-    width: 0,
-    height: 15,
-  },
-  elevation: 20,
-  shadowRadius: 30,
-  shadowOpacity: 1,
-  borderWidth: 1,
-  width: '90%',
-  paddingTop: 7,
-  paddingBottom: 5,
-  borderRadius: 20,
-};
 
 const LoadingOverlay = ({
   title,
@@ -34,13 +15,8 @@ const LoadingOverlay = ({
   title: string;
   subTitle?: string;
 }) => (
-  <CenteredContainer style={styles.overlayWrapper}>
-    <CenteredContainer
-      backgroundColor="white"
-      borderColor="whiteOverlay"
-      shadowColor="overlay"
-      {...overlayStyle}
-    >
+  <OverlayContainer>
+    <CenteredContainer paddingTop={2}>
       <ActivityIndicator color={colors.blueText} />
       <Text color="black" marginTop={5} fontSize={18} weight="bold">
         {title}
@@ -51,7 +27,7 @@ const LoadingOverlay = ({
         </Text>
       )}
     </CenteredContainer>
-  </CenteredContainer>
+  </OverlayContainer>
 );
 
 export default memo(LoadingOverlay);
