+++ cardstack/src/components/OverlayContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as OverlayContainer } from './OverlayContainer';
