+++ cardstack/src/components/SeedPhraseTable/SeedPhraseTable.tsx
@@ -50,7 +50,7 @@ export const SeedPhraseTable = ({
     new Animated.Value(!hideOnOpen ? animConfig.visible : animConfig.blurred)
   ).current;
 
-  const { copyToClipboard } = useCopyToast({});
+  const { copyToClipboard } = useCopyToast();
 
   const wordColumns = useMemo(
     () =>
