+++ cardstack/src/components/index.ts
@@ -35,7 +35,9 @@ export * from './TransactionList';
 export * from './TransactionConfirmationSheet';
 export * from './SafeAreaView';
 export * from './FilterOption';
+export * from './OverlayContainer';
 export * from './LoadingOverlay';
+export * from './ToastOverlay';
 export * from './BalanceSection';
 export * from './AnimatedPressable';
 export * from './CardPressable';
@@ -57,4 +59,3 @@ export * from './BackupStatus';
 export * from './PageWithStackHeader';
 export * from './SeedPhraseTable';
 export * from './WordPressable';
-export * from './ToastOverlay';
